refactor: Move the Hive-specific properties from common to presto-hive
This change moves session properties and config properties about
`orc_use_column_names` and `parquet_use_column_names` from
module `presto-hive-common` to `presto-hive`.

Unlike Hive, which relies on column order for schema matching by
default, modern lakehouse formats do not need the configuration for
column-name-based mapping to handle certain kinds of schema evolution.
Therefore, these two properties are unique to the Hive connector and
should not be placed in `presto-hive-common`, as doing so would
unnecessarily expose them to other lakehouse connectors.
