+++ presto-hive-common/src/main/java/com/facebook/presto/hive/HiveCommonClientConfig.java
@@ -37,14 +37,12 @@ public class HiveCommonClientConfig
     private DataSize orcStreamBufferSize = new DataSize(8, MEGABYTE);
     private OrcWriteValidationMode orcWriterValidationMode = OrcWriteValidationMode.BOTH;
     private double orcWriterValidationPercentage;
-    private boolean useOrcColumnNames;
     private DataSize orcTinyStripeThreshold = new DataSize(8, MEGABYTE);
     private boolean parquetBatchReadOptimizationEnabled;
     private boolean parquetEnableBatchReaderVerification;
     private DataSize parquetMaxReadBlockSize = new DataSize(16, MEGABYTE);
     private boolean rangeFiltersOnSubscriptsEnabled;
     private boolean readNullMaskedParquetEncryptedValueEnabled;
-    private boolean useParquetColumnNames;
     private boolean zstdJniDecompressionEnabled;
     private String catalogName;
     private DataSize affinitySchedulingFileSectionSize = new DataSize(256, MEGABYTE);
@@ -184,19 +182,6 @@ public class HiveCommonClientConfig
         return this;
     }
 
-    public boolean isUseOrcColumnNames()
-    {
-        return useOrcColumnNames;
-    }
-
-    @Config("hive.orc.use-column-names")
-    @ConfigDescription("Access ORC columns using names from the file first, and fallback to Hive schema column names if not found to ensure backward compatibility with old data")
-    public HiveCommonClientConfig setUseOrcColumnNames(boolean useOrcColumnNames)
-    {
-        this.useOrcColumnNames = useOrcColumnNames;
-        return this;
-    }
-
     @NotNull
     public DataSize getOrcTinyStripeThreshold()
     {
@@ -273,19 +258,6 @@ public class HiveCommonClientConfig
     public boolean getReadNullMaskedParquetEncryptedValue()
     {
         return this.readNullMaskedParquetEncryptedValueEnabled;
-    }
-
-    public boolean isUseParquetColumnNames()
-    {
-        return useParquetColumnNames;
-    }
-
-    @Config("hive.parquet.use-column-names")
-    @ConfigDescription("Access Parquet columns using names from the file")
-    public HiveCommonClientConfig setUseParquetColumnNames(boolean useParquetColumnNames)
-    {
-        this.useParquetColumnNames = useParquetColumnNames;
-        return this;
     }
 
     public boolean isZstdJniDecompressionEnabled()
