+++ presto-hive-common/src/main/java/com/facebook/presto/hive/HiveCommonSessionProperties.java
@@ -42,8 +42,6 @@ public class HiveCommonSessionProperties
     public static final String RANGE_FILTERS_ON_SUBSCRIPTS_ENABLED = "range_filters_on_subscripts_enabled";
     @VisibleForTesting
     public static final String PARQUET_BATCH_READ_OPTIMIZATION_ENABLED = "parquet_batch_read_optimization_enabled";
-    @VisibleForTesting
-    public static final String ORC_USE_COLUMN_NAMES = "orc_use_column_names";
 
     public static final String NODE_SELECTION_STRATEGY = "node_selection_strategy";
     private static final String ORC_BLOOM_FILTERS_ENABLED = "orc_bloom_filters_enabled";
@@ -60,7 +58,7 @@ public class HiveCommonSessionProperties
     private static final String ORC_ZSTD_JNI_DECOMPRESSION_ENABLED = "orc_zstd_jni_decompression_enabled";
     private static final String PARQUET_BATCH_READER_VERIFICATION_ENABLED = "parquet_batch_reader_verification_enabled";
     private static final String PARQUET_MAX_READ_BLOCK_SIZE = "parquet_max_read_block_size";
-    private static final String PARQUET_USE_COLUMN_NAMES = "parquet_use_column_names";
+
     public static final String READ_MASKED_VALUE_ENABLED = "read_null_masked_parquet_encrypted_value_enabled";
     public static final String AFFINITY_SCHEDULING_FILE_SECTION_SIZE = "affinity_scheduling_file_section_size";
     private final List<PropertyMetadata<?>> sessionProperties;
@@ -156,11 +154,6 @@ public class HiveCommonSessionProperties
                         hiveCommonClientConfig.isZstdJniDecompressionEnabled(),
                         true),
                 booleanProperty(
-                        ORC_USE_COLUMN_NAMES,
-                        "Access ORC columns using names from the file first, and fallback to Hive schema column names if not found to ensure backward compatibility with old data",
-                        hiveCommonClientConfig.isUseOrcColumnNames(),
-                        false),
-                booleanProperty(
                         PARQUET_BATCH_READ_OPTIMIZATION_ENABLED,
                         "Is Parquet batch read optimization enabled",
                         hiveCommonClientConfig.isParquetBatchReadOptimizationEnabled(),
@@ -174,11 +167,6 @@ public class HiveCommonSessionProperties
                         PARQUET_MAX_READ_BLOCK_SIZE,
                         "Parquet: Maximum size of a block to read",
                         hiveCommonClientConfig.getParquetMaxReadBlockSize(),
-                        false),
-                booleanProperty(
-                        PARQUET_USE_COLUMN_NAMES,
-                        "Experimental: Parquet: Access Parquet columns using names from the file",
-                        hiveCommonClientConfig.isUseParquetColumnNames(),
                         false),
                 booleanProperty(
                         READ_MASKED_VALUE_ENABLED,
@@ -269,11 +257,6 @@ public class HiveCommonSessionProperties
         return session.getProperty(ORC_ZSTD_JNI_DECOMPRESSION_ENABLED, Boolean.class);
     }
 
-    public static boolean isUseOrcColumnNames(ConnectorSession session)
-    {
-        return session.getProperty(ORC_USE_COLUMN_NAMES, Boolean.class);
-    }
-
     public static boolean isParquetBatchReadsEnabled(ConnectorSession session)
     {
         return session.getProperty(PARQUET_BATCH_READ_OPTIMIZATION_ENABLED, Boolean.class);
@@ -287,11 +270,6 @@ public class HiveCommonSessionProperties
     public static DataSize getParquetMaxReadBlockSize(ConnectorSession session)
     {
         return session.getProperty(PARQUET_MAX_READ_BLOCK_SIZE, DataSize.class);
-    }
-
-    public static boolean isUseParquetColumnNames(ConnectorSession session)
-    {
-        return session.getProperty(PARQUET_USE_COLUMN_NAMES, Boolean.class);
     }
 
     public static boolean isRangeFiltersOnSubscriptsEnabled(ConnectorSession session)
