+++ presto-hive-common/src/test/java/com/facebook/presto/hive/TestHiveCommonClientConfig.java
@@ -33,7 +33,6 @@ public class TestHiveCommonClientConfig
         ConfigAssertions.assertRecordedDefaults(ConfigAssertions.recordDefaults(HiveCommonClientConfig.class)
                 .setRangeFiltersOnSubscriptsEnabled(false)
                 .setNodeSelectionStrategy(NodeSelectionStrategy.valueOf("NO_PREFERENCE"))
-                .setUseParquetColumnNames(false)
                 .setParquetMaxReadBlockSize(new DataSize(16, DataSize.Unit.MEGABYTE))
                 .setOrcBloomFiltersEnabled(false)
                 .setOrcMaxMergeDistance(new DataSize(1, DataSize.Unit.MEGABYTE))
@@ -45,7 +44,6 @@ public class TestHiveCommonClientConfig
                 .setOrcOptimizedWriterEnabled(true)
                 .setOrcWriterValidationPercentage(0.0)
                 .setOrcWriterValidationMode(OrcWriteValidation.OrcWriteValidationMode.BOTH)
-                .setUseOrcColumnNames(false)
                 .setZstdJniDecompressionEnabled(false)
                 .setParquetBatchReaderVerificationEnabled(false)
                 .setParquetBatchReadOptimizationEnabled(false)
@@ -60,7 +58,6 @@ public class TestHiveCommonClientConfig
         Map<String, String> properties = new ImmutableMap.Builder<String, String>()
                 .put("hive.range-filters-on-subscripts-enabled", "true")
                 .put("hive.node-selection-strategy", "HARD_AFFINITY")
-                .put("hive.parquet.use-column-names", "true")
                 .put("hive.parquet.max-read-block-size", "66kB")
                 .put("hive.orc.bloom-filters.enabled", "true")
                 .put("hive.orc.max-merge-distance", "22kB")
@@ -72,7 +69,6 @@ public class TestHiveCommonClientConfig
                 .put("hive.orc.optimized-writer.enabled", "false")
                 .put("hive.orc.writer.validation-percentage", "0.16")
                 .put("hive.orc.writer.validation-mode", "DETAILED")
-                .put("hive.orc.use-column-names", "true")
                 .put("hive.zstd-jni-decompression-enabled", "true")
                 .put("hive.enable-parquet-batch-reader-verification", "true")
                 .put("hive.parquet-batch-read-optimization-enabled", "true")
@@ -84,7 +80,6 @@ public class TestHiveCommonClientConfig
         HiveCommonClientConfig expected = new HiveCommonClientConfig()
                 .setRangeFiltersOnSubscriptsEnabled(true)
                 .setNodeSelectionStrategy(HARD_AFFINITY)
-                .setUseParquetColumnNames(true)
                 .setParquetMaxReadBlockSize(new DataSize(66, DataSize.Unit.KILOBYTE))
                 .setOrcBloomFiltersEnabled(true)
                 .setOrcMaxMergeDistance(new DataSize(22, DataSize.Unit.KILOBYTE))
@@ -96,7 +91,6 @@ public class TestHiveCommonClientConfig
                 .setOrcOptimizedWriterEnabled(false)
                 .setOrcWriterValidationPercentage(0.16)
                 .setOrcWriterValidationMode(OrcWriteValidation.OrcWriteValidationMode.DETAILED)
-                .setUseOrcColumnNames(true)
                 .setZstdJniDecompressionEnabled(true)
                 .setParquetBatchReaderVerificationEnabled(true)
                 .setParquetBatchReadOptimizationEnabled(true)
